requirements.txt
# ── Core Agent Framework ────────────────────────────────
langchain-aws>=1.4.0
langchain-core>=1.2.0
langgraph>=1.1.0
langgraph-prebuilt>=1.0.0

# ── AWS / Bedrock ───────────────────────────────────────
boto3>=1.42.0

# ── Web Framework ──────────────────────────────────────
flask>=3.1.0

# ── HTTP & Config ──────────────────────────────────────
requests>=2.32.0
python-dotenv>=1.0.0

# ── Markdown rendering (server-side) ──────────────────
markdown>=3.7
